// src/app/models/recommended-intake.model.ts
export interface RecommendedIntake {
    daily_caloric_needs: number;
    carbohydrate: number;
    total_fiber: number;
    protein: number;
    fat: number;
    saturated_fatty_acids: number;
    trans_fatty_acids: number;
    alpha_linolenic_acid: number;
    linoleic_acid: number;
    dietary_cholesterol: number;
    total_water: number;
  }
  